requirements.txt
tensorflow-cpu
pyqt5
pillow
opencv-python
matplotlib